Add blocklist functionality
This adds blocklist functionality, while preserving the allowlist.
To make this more useful, we also introduce a default policy:  if an IP
matches neither allow- or block-list, it will be allowed or not based on
the default policy.

For CIDR-based lists, higher specificity CIDRs should always take
priority to lower-specificity ones.  In the event of a tie, the block
list should always take priority over the allow list.